[docs] Init DocsPage
