+++ packages/docs/src/app/components/AppHeader/AppHeaderLogo/AppHeaderLogo.tsx
@@ -1,10 +1,11 @@
 import GPTTurboLogo from "@components/GPTTurboLogo/GPTTurboLogo";
 import { Box, Group } from "@mantine/core";
+import Link from "next/link";
 
 const AppHeaderLogo = () => {
     return (
         <Group h="100%" w="100%">
-            <Box h="50%">
+            <Box component={Link} href="/" h="50%">
                 <GPTTurboLogo />
             </Box>
         </Group>
