+++ packages/docs/src/app/docs/page.tsx
@@ -1,0 +1,11 @@
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+    title: "Docs | GPT Turbo",
+};
+
+const DocsPage = () => {
+    return <div>DocsPage</div>;
+};
+
+export default DocsPage;
